TeamG_PatientRegistrationSystem/requirements.txt
astroid==2.0.4
colorama==0.4.0
diff-match-patch==20121119
Django==2.1.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-ranged-response==0.2.0
django-reversion==3.0.1
django-simple-captcha==0.5.9
et-xmlfile==1.0.1
future==0.17.0
httplib2==0.9.2
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.5.9
Pillow==5.3.0
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.6
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
visitor==0.1.3
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0